Try to fix func tests on travis

+++ test/functional/config/setup.js
@@ -14,5 +14,9 @@ beforeEach(function (done) {
 })
 
 afterEach(function (done) {
-  mongoose.connection.db.dropDatabase(done)
+  if (mongoose.connection.db) {
+    mongoose.connection.db.dropDatabase(done)
+  } else {
+    done()
+  }
 })
